pyproject.toml
[tool.poetry]
name = "deepterminal"
version = "0.1.0"
description = "Terminal-based trading application"
authors = ["Akas Antony antony.akas@gmail.com"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
textual = "^0.47.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
websocket-client = "^1.6.4"
pydantic = "^2.5.0"
rich = "^13.6.0"
pandas = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"

[tool.poetry.scripts]
trading-app = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"